--- contracts/button/mod.rs
@@ -1,3 +1,0 @@
-mod button;
-
-pub use button::*;
